Save system

+++ Projet Horde Survivor/Assets/Scripts/ShopManager.cs
@@ -6,7 +6,7 @@ public class ShopManager : MonoBehaviour
 {
     [SerializeField] private GameObject autoClicker;
     [SerializeField] private List<GameObject> autoClickers;
-    [SerializeField] private int price;
+    [SerializeField] private ulong price;
     [SerializeField] private int autoClickerLevel;
     
     public void Buy(int upgradeID)
